perm-checkify

+++ src/BaseClient/ClientHelperModules/requestHandler/guilds/deleteAutoModerationRule.ts
@@ -3,6 +3,9 @@ import error from '../../error.js';
 import { API } from '../../../Client.js';
 import cache from '../../cache.js';
 
+import getBotMemberFromGuild from '../../getBotMemberFromGuild.js';
+import requestHandlerError from '../../requestHandlerError.js';
+
 /**
  * Deletes an auto-moderation rule from a guild.
  * @param guild - The guild to delete the auto-moderation rule from.
@@ -11,10 +14,28 @@ import cache from '../../cache.js';
  * @returns A promise that resolves with the deleted auto-moderation rule,
  * or rejects with an error.
  */
-export default async (guild: Discord.Guild, ruleId: string, reason?: string) =>
- (cache.apis.get(guild.id) ?? API).guilds
+export default async (guild: Discord.Guild, ruleId: string, reason?: string) => {
+ if (!canDeleteAutoModerationRule(await getBotMemberFromGuild(guild))) {
+  const e = requestHandlerError(
+   `Cannot delete auto-moderation rule ${ruleId} in ${guild.name} / ${guild.id}`,
+   [Discord.PermissionFlagsBits.ManageGuild],
+  );
+
+  error(guild, e);
+  return e;
+ }
+
+ return (cache.apis.get(guild.id) ?? API).guilds
   .deleteAutoModerationRule(guild.id, ruleId, { reason })
   .catch((e) => {
    error(guild, new Error((e as Discord.DiscordAPIError).message));
    return e as Discord.DiscordAPIError;
   });
+};
+/**
+ * Checks if the given Discord GuildMember has the permission to delete an auto-moderation rule.
+ * @param me - The Discord GuildMember object representing the bot itself.
+ * @returns A boolean indicating whether the GuildMember has the required permission.
+ */
+export const canDeleteAutoModerationRule = (me: Discord.GuildMember) =>
+ me.permissions.has(Discord.PermissionFlagsBits.ManageGuild);
